+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.4.4 - 9 Feb 2025
+Bug fix:
+- [#55](https://github.com/elysiajs/node/pull/55) allow node cluster mode as default
+- createResponseHandler with more precision headers merging
+- simplify ws hook
+
 # 1.4.3 - 5 Jan 2025
 Bug fix:
 - [#54](https://github.com/elysiajs/node/issues/54) update srvx to 0.10.0